Move token generation into its own module
Move token generation into its own module

Test for TokenGenerator
